+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 
 ## 2.26 branch
 ### 2.26.8
+* javascript functions are now "camelCase": `tag_exist()` becoms `tagExists()`;
+    `add_tags()` becomes `addTags()`; `remove_tag()` becomes `removeTag()`;
+    `update_output_text()` becomes `updateOutputText()`;
+* many small fixes;
 * updated for `me-tools` `2.18.12`.
 
 ### 2.26.7